+++ src/tests/functional/orga/test_settings.py
@@ -17,7 +17,7 @@ def test_edit_mail_settings(orga_client, event, availability):
         follow=True,
         data={
             'mail_from': 'foo@bar.com',
-            'smtp_host': 'foo.bar.com',
+            'smtp_host': 'localhost',
             'smtp_password': '',
             'smtp_port': '25',
         }
@@ -26,6 +26,31 @@ def test_edit_mail_settings(orga_client, event, availability):
     event = Event.objects.get(pk=event.pk)
     assert event.settings.mail_from == 'foo@bar.com'
     assert event.settings.smtp_port == 25
+
+
+@pytest.mark.django_db
+def test_fail_unencrypted_mail_settings(orga_client, event, availability):
+    assert event.settings.mail_from != 'foo@bar.com'
+    assert event.settings.smtp_port != '25'
+    response = orga_client.get(
+        event.orga_urls.mail_settings,
+        follow=True,
+    )
+    assert response.status_code == 200
+    response = orga_client.post(
+        event.orga_urls.mail_settings,
+        follow=True,
+        data={
+            'mail_from': 'foo@bar.com',
+            'smtp_host': 'foo.bar.com',
+            'smtp_password': '',
+            'smtp_port': '25',
+        }
+    )
+    assert response.status_code == 200
+    event = Event.objects.get(pk=event.pk)
+    assert event.settings.mail_from != 'foo@bar.com'
+    assert event.settings.smtp_port != 25
 
 
 @pytest.mark.django_db
@@ -42,7 +67,7 @@ def test_test_mail_settings(orga_client, event, availability):
         follow=True,
         data={
             'mail_from': 'foo@bar.com',
-            'smtp_host': 'foo.bar.com',
+            'smtp_host': 'localhost',
             'smtp_password': '',
             'smtp_port': '25',
             'smtp_use_custom': '1',